Insert and withdraw now working correctly

+++ Bank/ValidationAttributes/NotLessThanZero.cs
@@ -10,7 +10,7 @@ namespace Bank.ValidationAttributes
     {
         protected override ValidationResult IsValid(object value, ValidationContext validationContext)
         {
-            if (int.Parse((value.ToString())) <= 0)
+            if (decimal.Parse((value.ToString())) <= 0)
             {
                 return new ValidationResult("Value must be greater than zero");
             }
